backend-python/requirements.txt
flask==1.1.2
flask-cors==3.0.9
pandas==1.0.1
numpy==1.19.1 
python-dotenv==0.14.0
gunicorn==20.0.4